Cast post_author to int before building author URLs
get_post_field() returns a string; the PHPDoc-type PHPStan lane fails
get_author_posts_url() and get_author_feed_link() on it. The pre-commit
PHPStan hook also gets the Makefile's 512M memory limit, since the default
128M crashed its worker on every commit.

+++ varnish-http-purge.php
@@ -2544,7 +2544,7 @@ class VarnishPurger {
 				// Pages and Woo Things don't need all this.
 				if ( $this_post_type && 'post' === $this_post_type ) {
 					// Author URLs:
-					$author_id = get_post_field( 'post_author', $post_id );
+					$author_id = (int) get_post_field( 'post_author', $post_id );
 					array_push(
 						$listofurls,
 						get_author_posts_url( $author_id ),
